Update index.html
Fixed Italics for "Vibe": I added a new CSS class called .vibe-italics and applied it to the word "Vibe" in the vibe output text. This ensures that only "Vibe" is italicized.

Fixed Save This Vibe Functionality: I modified the saveVibe() function to make sure it works correctly using html2canvas. This will now properly generate the screenshot of the vibe output and let the user download it.